# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Context-driven design prompt generation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Session {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateSessionBody {
  title: string;
  /** @nullable */
  description?: string | null;
}

export interface UpdateSessionBody {
  title?: string;
  /** @nullable */
  description?: string | null;
}

export type ContextItemType =
  (typeof ContextItemType)[keyof typeof ContextItemType];

export const ContextItemType = {
  note: "note",
  file: "file",
  image: "image",
  requirement: "requirement",
  paste: "paste",
} as const;

export interface ContextItem {
  id: number;
  sessionId: number;
  type: ContextItemType;
  /** @nullable */
  label?: string | null;
  content: string;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  filename?: string | null;
  /** @nullable */
  mimeType?: string | null;
  createdAt: string;
}

export type CreateContextItemBodyType =
  (typeof CreateContextItemBodyType)[keyof typeof CreateContextItemBodyType];

export const CreateContextItemBodyType = {
  note: "note",
  file: "file",
  image: "image",
  requirement: "requirement",
  paste: "paste",
} as const;

export interface CreateContextItemBody {
  type: CreateContextItemBodyType;
  /** @nullable */
  label?: string | null;
  content: string;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  filename?: string | null;
  /** @nullable */
  mimeType?: string | null;
}

export interface UpdateContextItemBody {
  /** @nullable */
  label?: string | null;
  content?: string;
}

export interface GeneratedPrompt {
  id: number;
  sessionId: number;
  content: string;
  /** @nullable */
  instruction?: string | null;
  version: number;
  createdAt: string;
  updatedAt: string;
}

export interface GeneratePromptBody {
  /**
   * Optional user instruction to guide the prompt generation
   * @nullable
   */
  instruction?: string | null;
}

export interface UpdatePromptBody {
  content?: string;
}

/**
 * Transcription provider to use
 */
export type TranscribeAudioBodyProvider =
  (typeof TranscribeAudioBodyProvider)[keyof typeof TranscribeAudioBodyProvider];

export const TranscribeAudioBodyProvider = {
  deepgram: "deepgram",
  openai: "openai",
} as const;

export interface TranscribeAudioBody {
  /** Base64-encoded audio data */
  audio: string;
  /** Audio MIME type (e.g. audio/webm) */
  mimetype?: string;
  /** Transcription provider to use */
  provider?: TranscribeAudioBodyProvider;
  /** Language code (e.g. da, en) */
  language?: string;
}

export interface TranscriptionResult {
  transcript: string;
}

export type SessionSummaryContextByType = { [key: string]: number };

export interface SessionSummary {
  sessionId: number;
  contextCount: number;
  transcriptSegmentCount: number;
  promptCount: number;
  contextByType: SessionSummaryContextByType;
  /** @nullable */
  latestPromptPreview?: string | null;
}

export interface TranscriptSegment {
  id: number;
  sessionId: number;
  speaker: string;
  text: string;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  recordingId?: string | null;
  createdAt: string;
}

export interface CreateTranscriptSegmentBody {
  speaker: string;
  text: string;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  recordingId?: string | null;
}

export interface CreateTranscriptSegmentsBatchBody {
  segments: CreateTranscriptSegmentBody[];
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface ErrorEnvelope {
  error: string;
}