# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Context-driven design prompt generation API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: sessions
    description: Session management
  - name: context
    description: Session context items
  - name: prompts
    description: Prompt generation and history
  - name: transcripts
    description: Session transcript segments from recordings
  - name: transcription
    description: Audio transcription via Deepgram
  - name: Storage
    description: Object storage upload and serving endpoints.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /sessions:
    get:
      operationId: listSessions
      tags: [sessions]
      summary: List all sessions
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
    post:
      operationId: createSession
      tags: [sessions]
      summary: Create a new session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionBody"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"

  /sessions/{id}:
    get:
      operationId: getSession
      tags: [sessions]
      summary: Get a session by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "404":
          description: Not found
    patch:
      operationId: updateSession
      tags: [sessions]
      summary: Update a session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSessionBody"
      responses:
        "200":
          description: Updated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "404":
          description: Not found
    delete:
      operationId: deleteSession
      tags: [sessions]
      summary: Delete a session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /sessions/{sessionId}/context:
    get:
      operationId: listContextItems
      tags: [context]
      summary: List context items for a session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of context items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContextItem"
    post:
      operationId: addContextItem
      tags: [context]
      summary: Add a context item to a session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContextItemBody"
      responses:
        "201":
          description: Created context item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContextItem"

  /sessions/{sessionId}/context/{id}:
    patch:
      operationId: updateContextItem
      tags: [context]
      summary: Update a context item
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContextItemBody"
      responses:
        "200":
          description: Updated context item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContextItem"
        "404":
          description: Not found
    delete:
      operationId: deleteContextItem
      tags: [context]
      summary: Delete a context item
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /sessions/{sessionId}/prompts:
    get:
      operationId: listPrompts
      tags: [prompts]
      summary: List generated prompts for a session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of prompts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeneratedPrompt"
    post:
      operationId: generatePrompt
      tags: [prompts]
      summary: Generate a new prompt from current session context
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneratePromptBody"
      responses:
        "201":
          description: Generated prompt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedPrompt"
        "400":
          description: No context available

  /sessions/{sessionId}/prompts/{id}:
    patch:
      operationId: updatePrompt
      tags: [prompts]
      summary: Update (edit) a generated prompt
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePromptBody"
      responses:
        "200":
          description: Updated prompt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedPrompt"
        "404":
          description: Not found
    delete:
      operationId: deletePrompt
      tags: [prompts]
      summary: Delete a prompt
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /sessions/{sessionId}/transcripts:
    get:
      operationId: listTranscriptSegments
      tags: [transcripts]
      summary: List transcript segments for a session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of transcript segments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TranscriptSegment"
    post:
      operationId: addTranscriptSegment
      tags: [transcripts]
      summary: Add a single transcript segment
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTranscriptSegmentBody"
      responses:
        "201":
          description: Created segment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptSegment"
    delete:
      operationId: clearTranscriptSegments
      tags: [transcripts]
      summary: Delete all transcript segments for a session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /sessions/{sessionId}/transcripts/batch:
    post:
      operationId: addTranscriptSegmentsBatch
      tags: [transcripts]
      summary: Add multiple transcript segments at once
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTranscriptSegmentsBatchBody"
      responses:
        "201":
          description: Created segments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TranscriptSegment"

  /sessions/{sessionId}/transcripts/{id}:
    delete:
      operationId: deleteTranscriptSegment
      tags: [transcripts]
      summary: Delete a single transcript segment
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /transcribe:
    post:
      operationId: transcribeAudio
      tags: [transcription]
      summary: Transcribe audio using selected provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscribeAudioBody"
      responses:
        "200":
          description: Transcription result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionResult"
        "400":
          description: Missing audio data

  /sessions/{sessionId}/summary:
    get:
      operationId: getSessionSummary
      tags: [sessions]
      summary: Get a quick summary of session context stats
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Session summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionSummary"

  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /storage/public-objects/{filePath}:
    get:
      tags: [Storage]
      operationId: getPublicObject
      summary: Serve a public asset from PUBLIC_OBJECT_SEARCH_PATHS
      parameters:
        - name: filePath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed with correct Content-Type.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: File not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed with correct Content-Type.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Session:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
        - updatedAt

    CreateSessionBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
      required:
        - title

    UpdateSessionBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]

    ContextItem:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        type:
          type: string
          enum: [note, file, image, requirement, paste]
        label:
          type: ["string", "null"]
        content:
          type: string
        fileUrl:
          type: ["string", "null"]
        filename:
          type: ["string", "null"]
        mimeType:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - type
        - content
        - createdAt

    CreateContextItemBody:
      type: object
      properties:
        type:
          type: string
          enum: [note, file, image, requirement, paste]
        label:
          type: ["string", "null"]
        content:
          type: string
        fileUrl:
          type: ["string", "null"]
        filename:
          type: ["string", "null"]
        mimeType:
          type: ["string", "null"]
      required:
        - type
        - content

    UpdateContextItemBody:
      type: object
      properties:
        label:
          type: ["string", "null"]
        content:
          type: string

    GeneratedPrompt:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        content:
          type: string
        instruction:
          type: ["string", "null"]
        version:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - content
        - version
        - createdAt
        - updatedAt

    GeneratePromptBody:
      type: object
      properties:
        instruction:
          type: ["string", "null"]
          description: Optional user instruction to guide the prompt generation

    UpdatePromptBody:
      type: object
      properties:
        content:
          type: string

    TranscribeAudioBody:
      type: object
      properties:
        audio:
          type: string
          description: Base64-encoded audio data
        mimetype:
          type: string
          description: Audio MIME type (e.g. audio/webm)
        provider:
          type: string
          enum: [deepgram, openai]
          description: Transcription provider to use
        language:
          type: string
          description: Language code (e.g. da, en)
      required:
        - audio

    TranscriptionResult:
      type: object
      properties:
        transcript:
          type: string
      required:
        - transcript

    SessionSummary:
      type: object
      properties:
        sessionId:
          type: integer
        contextCount:
          type: integer
        transcriptSegmentCount:
          type: integer
        promptCount:
          type: integer
        contextByType:
          type: object
          additionalProperties:
            type: integer
        latestPromptPreview:
          type: ["string", "null"]
      required:
        - sessionId
        - contextCount
        - transcriptSegmentCount
        - promptCount
        - contextByType

    TranscriptSegment:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        speaker:
          type: string
        text:
          type: string
        language:
          type: ["string", "null"]
        recordingId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - speaker
        - text
        - createdAt

    CreateTranscriptSegmentBody:
      type: object
      properties:
        speaker:
          type: string
        text:
          type: string
        language:
          type: ["string", "null"]
        recordingId:
          type: ["string", "null"]
      required:
        - speaker
        - text

    CreateTranscriptSegmentsBatchBody:
      type: object
      properties:
        segments:
          type: array
          items:
            $ref: "#/components/schemas/CreateTranscriptSegmentBody"
      required:
        - segments

    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1

    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'

    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string