# lib/api-zod/src/generated/types/createTranscriptSegmentsBatchBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Context-driven design prompt generation API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTranscriptSegmentBody } from "./createTranscriptSegmentBody";

export interface CreateTranscriptSegmentsBatchBody {
  segments: CreateTranscriptSegmentBody[];
}