# lib/api-zod/src/generated/types/sessionSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Context-driven design prompt generation API
 * OpenAPI spec version: 0.1.0
 */
import type { SessionSummaryContextByType } from "./sessionSummaryContextByType";

export interface SessionSummary {
  sessionId: number;
  contextCount: number;
  transcriptSegmentCount: number;
  promptCount: number;
  contextByType: SessionSummaryContextByType;
  /** @nullable */
  latestPromptPreview?: string | null;
}