# lib/api-zod/src/generated/types/transcriptSegment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Context-driven design prompt generation API
 * OpenAPI spec version: 0.1.0
 */

export interface TranscriptSegment {
  id: number;
  sessionId: number;
  speaker: string;
  text: string;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  recordingId?: string | null;
  createdAt: Date;
}